requirements.txt
python-docx>=0.8.11
requests>=2.28.0
pytest>=7.0.0  # Optional
